[UPDATE] 회원가입 기능 업데이트(Valid코드, 패키지명 변경, 테스트코드 수정 및 ,코드 리팩터링)
