--- src/main/java/com/ssg/usms/business/User/annotation/CustomUsername.java
@@ -1,21 +1,0 @@
-package com.ssg.usms.business.User.annotation;
-
-import javax.validation.Constraint;
-import javax.validation.Payload;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-
-@Target(ElementType.FIELD)
-@Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = usernameValidator.class)
-public @interface CustomUsername {
-
-    String message() default "유효하지 않은 아이디 형식입니다.";
-
-    Class<?>[] groups() default {};
-
-    Class<? extends Payload>[] payload() default {};
-}
