--- src/main/java/com/ssg/usms/business/User/exception/NotAllowedEmailFormException.java
@@ -1,12 +1,0 @@
-package com.ssg.usms.business.User.exception;
-
-import static com.ssg.usms.business.constant.CustomStatusCode.NOT_ALLOWED_EMAIL_FORM;
-
-
-public class NotAllowedEmailFormException extends IllegalSignUpFormException{
-
-    public NotAllowedEmailFormException(String message){
-
-        super(NOT_ALLOWED_EMAIL_FORM, message);
-    }
-}
