--- src/main/java/com/ssg/usms/business/User/exception/NotAllowedKeyExcetpion.java
@@ -1,12 +1,0 @@
-package com.ssg.usms.business.User.exception;
-
-import static com.ssg.usms.business.constant.CustomStatusCode.NOT_ALLOWED_KEY;
-
-public class NotAllowedKeyExcetpion extends IllegalSignUpFormException{
-
-    public NotAllowedKeyExcetpion(String message){
-
-        super(NOT_ALLOWED_KEY, message);
-    }
-
-}
