--- src/main/java/com/ssg/usms/business/User/persistence/UserRole.java
@@ -1,5 +1,0 @@
-package com.ssg.usms.business.User.persistence;
-
-public enum UserRole {
-    ADMIN,STORE_OWNER;
-}
