+++ src/main/java/com/ssg/usms/business/config/SecurityConfig.java
@@ -1,6 +1,9 @@
 package com.ssg.usms.business.config;
 
 
+import com.ssg.usms.business.Security.login.UsmsLoginConfiguer;
+import com.ssg.usms.business.Security.login.authority.UsmsAccessDeniedHandler;
+import com.ssg.usms.business.Security.login.authority.UsmsForbiddenEntryPoint;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpMethod;
@@ -9,6 +12,9 @@ import org.springframework.security.config.annotation.web.configuration.EnableWe
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.access.intercept.AuthorizationFilter;
+import org.springframework.security.web.authentication.AuthenticationFailureHandler;
+import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
+import org.springframework.security.web.authentication.logout.LogoutSuccessHandler;
 import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 import org.springframework.security.web.util.matcher.RequestMatcher;
 
@@ -22,16 +28,30 @@ public class SecurityConfig {
     }
 
     @Bean
-    public SecurityFilterChain defaultSecurityFilterChain(HttpSecurity http) throws Exception{
+    public SecurityFilterChain defaultSecurityFilterChain(HttpSecurity http, AuthenticationSuccessHandler loginSuccessHandler, AuthenticationFailureHandler loginFailureHandler, LogoutSuccessHandler logoutSuccessHandler) throws Exception{
 
         http
                 .authorizeHttpRequests((auth) -> auth
                         .antMatchers(HttpMethod.GET,("/api/users")).hasRole("ADMIN")
                         .antMatchers(HttpMethod.POST,("/api/users")).permitAll()
+                        .antMatchers(HttpMethod.POST,("/api/identification")).permitAll()
                         .anyRequest().authenticated());
-        //d임시
+
         http
                 .csrf().disable();
+
+        http.exceptionHandling()
+                        .accessDeniedHandler(new UsmsAccessDeniedHandler())
+                                .authenticationEntryPoint(new UsmsForbiddenEntryPoint());
+
+        http.apply(new UsmsLoginConfiguer<>())
+                .loginProcessingUrl("/api/login")
+                .successHandler(loginSuccessHandler)
+                .failureHandler(loginFailureHandler);
+
+        http.logout()
+                .logoutUrl("/api/logout")
+                .logoutSuccessHandler(logoutSuccessHandler);
