+++ src/main/java/com/ssg/usms/business/user/annotation/Email.java
@@ -1,0 +1,26 @@
+package com.ssg.usms.business.user.annotation;
+
+
+import javax.validation.Constraint;
+import javax.validation.Payload;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.FIELD)
+@Retention(RetentionPolicy.RUNTIME)
+@Constraint(validatedBy = EmailVaildator.class)
+public @interface Email {
+
+    String message() default "유효하지 않은 이메일 형식입니다.";
+
+    Class<?>[] groups() default {};
+
+    Class<? extends Payload>[] payload() default {};
+
+
+}
+
+
+
