+++ src/main/java/com/ssg/usms/business/user/annotation/NickName.java
@@ -1,0 +1,22 @@
+package com.ssg.usms.business.user.annotation;
+
+
+import javax.validation.Constraint;
+import javax.validation.Payload;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.FIELD)
+@Retention(RetentionPolicy.RUNTIME)
+@Constraint(validatedBy = NickNameValidator.class)
+public @interface NickName {
+
+    String message() default "닉네임은 1자이상 10자이하 특수문자를 포함할수없습니다.";
+
+    Class<?>[] groups() default {};
+
+    Class<? extends Payload>[] payload() default {};
+
+}
