+++ src/main/java/com/ssg/usms/business/user/exception/AlreadyExistPhoneNumException.java
@@ -1,0 +1,11 @@
+package com.ssg.usms.business.user.exception;
+
+import static com.ssg.usms.business.constant.CustomStatusCode.ALREADY_EXIST_PHONE_NUM;
+
+public class AlreadyExistPhoneNumException extends IllegalSignUpFormException{
+
+    public AlreadyExistPhoneNumException(String message){
+
+        super(ALREADY_EXIST_PHONE_NUM,message);
+    }
+}
