+++ src/main/java/com/ssg/usms/business/user/exception/IllegalSignUpFormException.java
@@ -1,0 +1,31 @@
+package com.ssg.usms.business.user.exception;
+
+
+import com.ssg.usms.business.error.ErrorResponseDto;
+import lombok.Getter;
+
+@Getter
+public abstract class IllegalSignUpFormException extends RuntimeException {
+
+    private final ErrorResponseDto errorResponseDto;
+
+    public IllegalSignUpFormException(String message){
+
+        errorResponseDto = new ErrorResponseDto(400, message);
+    }
+
+    public IllegalSignUpFormException(int code, String message){
+
+        errorResponseDto = new ErrorResponseDto(code, message);
+    }
+
+
+
+    @Override
+    public String getMessage(){
+
+        return errorResponseDto.getMessage();
+    }
+
+
+}
