+++ src/main/java/com/ssg/usms/business/user/persistence/UserRole.java
@@ -1,0 +1,15 @@
+package com.ssg.usms.business.user.persistence;
+
+import lombok.Getter;
+
+@Getter
+public enum UserRole {
+    ADMIN(0)
+    ,STORE_OWNER(1);
+
+    private final int code;
+
+    UserRole(int code){
+        this.code = code;
+    }
+}
