+++ src/test/java/com/ssg/usms/business/SignUp/SignUpServiceTest.java
@@ -1,12 +1,12 @@
 package com.ssg.usms.business.SignUp;
 
 
-import com.ssg.usms.business.User.Repository.UserRepository;
-import com.ssg.usms.business.User.exception.AlreadyExistIdException;
-import com.ssg.usms.business.User.exception.AlreadyExistPhoneNumException;
-import com.ssg.usms.business.User.persistence.HttpRequestSignUpDto;
-import com.ssg.usms.business.User.persistence.UsmsUser;
-import com.ssg.usms.business.User.service.SignUpService;
+import com.ssg.usms.business.user.Repository.UserRepository;
+import com.ssg.usms.business.user.exception.AlreadyExistIdException;
+import com.ssg.usms.business.user.exception.AlreadyExistPhoneNumException;
+import com.ssg.usms.business.user.persistence.HttpRequestSignUpDto;
+import com.ssg.usms.business.user.persistence.UsmsUser;
+import com.ssg.usms.business.user.service.SignUpService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -15,9 +15,11 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.test.context.ActiveProfiles;
 
+import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
+import static org.mockito.BDDMockito.willThrow;
 
 @ActiveProfiles("test")
 @SpringBootTest
@@ -31,70 +33,31 @@ public class SignUpServiceTest {
 
     @BeforeEach
     public void setup() {
-        service = new SignUpService(bCryptPasswordEncoder,repository);
+        service = new SignUpService(bCryptPasswordEncoder, repository);
     }
 
-    @DisplayName("이미 중복된 아이디로 회원가입을 요청한 경우 AlreadyExistId에러를 날린다.")
+    @DisplayName("중복된 아이디를 요청한 경우 AlreadyExistIdException 을 던진다.")
     @Test
-    public void testSinupWithAlreadyExistId(){
-//        given
-        given(repository.existsByUsername("aksenaksen")).willReturn(true);
+    public void testSinupWithDupicateId() {
 
-        assertThrows(AlreadyExistIdException.class,() -> service.CheckDuplicatedId("aksenaksen"));
+        given(repository.existsByUsername(any())).willReturn(true);
+
+        HttpRequestSignUpDto dto = new HttpRequestSignUpDto();
+        dto.setUsername("httpRequestSign");
+        dto.setPassword("hashedpassword123@");
+        dto.setPhoneNum("010-1234-24124");
+        dto.setEmail("asdf123@naer.com");
+        dto.setNickname("hihello");
+
+
+        assertThrows(AlreadyExistIdException.class, () -> service.SignUp(dto));
     }
-
-    @DisplayName("이미 중복된 전화번호로 회원가입을 요청한 경우 AlreadyExistPhoneNum에러를 날린다.")
-    @Test
-    public void testSinupWithAlreadyExistPhoneNum(){
-//        given
-        given(repository.existsByPhoneNumber("010-1234-2323")).willReturn(true);
-
-        assertThrows(AlreadyExistPhoneNumException.class,() -> service.CheckDuplicatePhoneNumber("010-1234-2323"));
-    }
-
-    @DisplayName("올바른 전화번호로 회원가입을 요청한 경우 duplicate 함수에서 false리턴")
-    @Test
-    public void testDuplicateWithValidPhoneNum(){
-//        given
-        given(repository.existsByPhoneNumber("010-1234-2323")).willReturn(false);
-        boolean result = service.CheckDuplicatePhoneNumber("010-1234-2323");
-
-        assertTrue(result);
-    }
-
-    @DisplayName("올바른 아이디로 회원가입을 요청한 경우 duplicate 함수에서 true리턴")
-    @Test
-    public void testDuplicateWithValidId(){
-//        given
-        given(repository.existsByUsername("aksenaksen")).willReturn(false);
-        boolean result = service.CheckDuplicatedId("aksenaksen");
-
-        assertTrue(result);
-    }
-//
-//
-@DisplayName("중복된 아이디를 요청한 경우 AlreadyExistIdException 을 던진다.")
-@Test
-public void testSinupWithDupicateId(){
-//        given
-    given(service.CheckDuplicatedId(any())).willThrow(new AlreadyExistIdException("이미 존재하는 아이디입니다."));
-
-    HttpRequestSignUpDto dto = new HttpRequestSignUpDto();
-    dto.setUsername("httpRequestSign");
-    dto.setPassword("hashedpassword123@");
-    dto.setPhoneNum("010-1234-24124");
-    dto.setEmail("asdf123@naer.com");
-    dto.setNickname("hihello");
-
-
-    assertThrows(AlreadyExistIdException.class, () -> service.SignUp(dto));
-}
 
     @DisplayName("중복된 전화번호를 요청한 경우 AlreadyExistPhoneNumbException 을 던진다..")
     @Test
-    public void testSinupWithDupicatePhoneNum(){
-//        given
-        given(service.CheckDuplicatePhoneNumber(any())).willThrow(new AlreadyExistPhoneNumException("이미 존재하는 전화번호입니다."));
+    public void testSinupWithDupicatePhoneNum() {
+
+        given(repository.existsByPhoneNumber(any())).willReturn(true);
 
         HttpRequestSignUpDto dto = new HttpRequestSignUpDto();
         dto.setUsername("httpRequestSign");
@@ -109,12 +72,9 @@ public void testSinupWithDupicateId(){
 
     @DisplayName("올바른 값을 요청한 경우 200 ResponseEntity를 반환한다.")
     @Test
-    public void testSinupWithValidParam(){
-//        given
-        given(service.CheckDuplicatePhoneNumber(any())).willReturn(false);
-        given(service.CheckDuplicatePhoneNumber(any())).willReturn(false);
+    public void testSinupWithValidParam() {
 
-        UsmsUser user= UsmsUser.builder()
+        UsmsUser user = UsmsUser.builder()
                 .username("httpRequestSign")
                 .password("hashedpassword123@")
                 .personName("hihello")
@@ -131,10 +91,7 @@ public void testSinupWithDupicateId(){
         dto.setEmail("asdf123@naer.com");
         dto.setNickname("hihello");
         // then
-        assertTrue(service.SignUp(dto));
+        assertThatCode(() -> service.SignUp(dto))
+                .doesNotThrowAnyException();
     }
-
-
-
-
 }
