+++ src/test/java/com/ssg/usms/business/SignUp/UserRepositoryTest.java
@@ -1,8 +1,8 @@
 package com.ssg.usms.business.SignUp;
 
 
-import com.ssg.usms.business.User.Repository.UserRepository;
-import com.ssg.usms.business.User.persistence.UsmsUser;
+import com.ssg.usms.business.user.Repository.UserRepository;
+import com.ssg.usms.business.user.persistence.UsmsUser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
